re-analysis for Reviewer 2 of Frontiers: remove entropy
Reviewer 2 complained that entropy wasn't properly describe in the paper,
but we still use it as regressor. To not confuse him (or anyone else)
further we decided to remove the entropy regressor. This is anyway only
relevant for Figure 4. I did not spot any changes in the results (except
for the permutations).